Add support for Required Bosses
There is a new section added to the settings popup that displays the
current Required Bosses setting and which bosses are required.

Non-required dungeons (and boss-related mail) will be hidden from the
tracker.

The logic for opening the Puppet Ganon door now accounts for being able
to defeat required bosses.